IGNITE-5000 (Rename Ignite Math module to Ignite ML module):   Split examples.
