+++ examples/pom-standalone-lgpl.xml
@@ -102,6 +102,19 @@
             <activation>
                 <jdk>[1.8,)</jdk>
             </activation>
+        </profile>
+
+        <profile>
+            <id>ml</id>
+
+            <properties>
+                <java8.folder>src/main/ml</java8.folder>
+                <java.ver>1.8</java.ver>
+            </properties>
+
+            <activation>
+                <jdk>[1.8,)</jdk>
+            </activation>
 
             <dependencies>
                 <dependency>
@@ -238,6 +251,7 @@
                                 <source>${lgpl.folder}</source>
                                 <source>${java8.folder}</source>
                                 <source>${spark.folder}</source>
+                                <source>${ml.folder}</source>
                             </sources>
                         </configuration>
                     </execution>
